refactor: rename mirror.py to pending_mirrors.py
'mirror' alone names neither what is mirrored nor that anything is pending;
the module owns the pending set + cursor, so it gets the heartbeat_state.py
naming shape. Concept name (pending-mirror drain) unchanged everywhere.
